Refine social login

+++ api/app/users/views/social_auth_views.py
@@ -103,7 +103,7 @@ class UserSocialDeleteAPI(generics.GenericAPIView):
     permission_classes = [IsAuthenticated]
     authentication_classes = [JWTAuthentication]
 
-    def delete(self, request, *args, **kwargs) -> Response:
+    def delete(self, request, *args, **kwargs):
         logger.info("DELETE /api/v1/users/social/delete")
         serializer = self.get_serializer(data=request.data)
         if serializer.is_valid():
@@ -111,11 +111,8 @@ class UserSocialDeleteAPI(generics.GenericAPIView):
                 email = serializer.validated_data["email"]
                 refresh_token = serializer.validated_data["refresh_token"]
                 oauth_platform = serializer.validated_data["oauth_platform"]
-                user = User.objects.get(email=email, oauth_platform=oauth_platform)
-
-                if not user.is_active:
-                    logger.warning(f"Inactive user attempted to delete account: {email}")
-                    return Response({"message": "User is inactive"}, status=status.HTTP_400_BAD_REQUEST)
+                user_oauth = UserOAuth2Platform.objects.get(user__email=email, oauth_platform=oauth_platform)
+                user = user_oauth.user
 
                 # RefreshToken을 블랙리스트에 추가
                 try:
@@ -130,8 +127,13 @@ class UserSocialDeleteAPI(generics.GenericAPIView):
                 # 사용자가 소유한 모든 OutstandingToken 삭제
                 OutstandingToken.objects.filter(user=user).delete()
 
-                # 사용자 삭제
-                user.delete()
+                # 소셜 계정 정보 삭제
+                user_oauth.delete()
+
+                # 만약 사용자가 더 이상 다른 소셜 계정이 없고, 일반 계정도 없다면 사용자 삭제
+                if not UserOAuth2Platform.objects.filter(user=user).exists():
+                    user.delete()
+
                 logger.info(f"User {email} deleted successfully.")
                 return Response({"message": "User deleted successfully"}, status=status.HTTP_200_OK)
             except User.DoesNotExist:
@@ -264,7 +266,7 @@ class UserNaverLoginCallBackAPI(generics.GenericAPIView):
     serializer_class = EmptySerializer
     permission_classes = [AllowAny]
 
-    def get(self, request, *args, **kwargs) -> Response:
+    def get(self, request, *args, **kwargs):
         # https://developers.naver.com/docs/login/api/api.md
         logger.info("GET /api/v1/users/oauth2/naver/callback")
 
@@ -333,9 +335,11 @@ class UserNaverLoginCallBackAPI(generics.GenericAPIView):
 
         # 사용자 존재 여부 확인해서 새로운 사용자를 생성하거나 기존 사용자가 있다면, 로그인
         try:
-            user_oauth2 = UserOAuth2Platform.objects.select_related('user').filter(
-                user__email=user_email, oauth_platform='naver'
-            ).first()
+            user_oauth2 = (
+                UserOAuth2Platform.objects.select_related("user")
+                .filter(user__email=user_email, oauth_platform="naver")
+                .first()
+            )
 
             if user_oauth2:
                 # 이미 존재하는 사용자의 경우
@@ -352,7 +356,7 @@ class UserNaverLoginCallBackAPI(generics.GenericAPIView):
                     last_login=timezone.now(),
                     is_active=True,
                 )
-                UserOAuth2Platform.objects.create(user=user, oauth_platform='naver', oauth2_user_id=user_id)
+                UserOAuth2Platform.objects.create(user=user, oauth_platform="naver", oauth2_user_id=user_id)
                 logger.info(f"New user created: {user_email}")
 
         except Exception as e:
@@ -370,7 +374,9 @@ class UserNaverLoginCallBackAPI(generics.GenericAPIView):
         tokens = get_jwt_tokens_for_user(user)
 
         logger.info(f"User logged in successfully: {user_email}")
-        return redirect(f"{os.getenv('MAIN_DOMAIN')}/auth/callback?access_token={tokens['access']}&refresh_token={tokens['refresh']}")
+        return redirect(
+            f"{os.getenv('MAIN_DOMAIN')}/auth/callback?access_token={tokens['access']}&refresh_token={tokens['refresh']}"
+        )
 
 
 @extend_schema(
@@ -557,9 +563,11 @@ class UserGoogleLoginCallBackAPI(generics.GenericAPIView):
             return Response({"message": "Failed to get user email and sub(id)"}, status=status.HTTP_400_BAD_REQUEST)
 
         try:
-            user_oauth2 = UserOAuth2Platform.objects.select_related('user').filter(
-                user__email=user_email, oauth_platform='google'
-            ).first()
+            user_oauth2 = (
+                UserOAuth2Platform.objects.select_related("user")
+                .filter(user__email=user_email, oauth_platform="google")
+                .first()
+            )
 
             if user_oauth2:
                 # 이미 존재하는 사용자의 경우
@@ -576,7 +584,7 @@ class UserGoogleLoginCallBackAPI(generics.GenericAPIView):
                     last_login=timezone.now(),
                     is_active=True,
                 )
-                UserOAuth2Platform.objects.create(user=user, oauth_platform='google', oauth2_user_id=user_id)
+                UserOAuth2Platform.objects.create(user=user, oauth_platform="google", oauth2_user_id=user_id)
                 logger.info(f"New user created: {user_email}")
 
         except Exception as e:
@@ -589,4 +597,6 @@ class UserGoogleLoginCallBackAPI(generics.GenericAPIView):
         tokens = get_jwt_tokens_for_user(user)
 
         logger.info(f"User logged in successfully: {user_email}")
-        return redirect(f"{os.getenv('MAIN_DOMAIN')}/auth/callback?access_token={tokens['access']}&refresh_token={tokens['refresh']}")
+        return redirect(
+            f"{os.getenv('MAIN_DOMAIN')}/auth/callback?access_token={tokens['access']}&refresh_token={tokens['refresh']}"
+        )
